requirements.txt
Django==4.1.1
djangorestframework==3.13.1
python-decouple==3.6
stripe==4.1.0
